requirements.txt
Django
djangorestframework
django-filter
django-simple-history
django-import-export
